requirements.txt
streamlit
opencv-python-headless
numpy
pandas
pydeck
psycopg2-binary